+++ lib/exceptions/resolve_token_exception.dart
@@ -1,0 +1,10 @@
+class ResolveTokenException {
+  final String message;
+
+  ResolveTokenException(this.message);
+
+  @override
+  String toString() {
+    return 'ResolveTokenException{message: $message}';
+  }
+}