+++ test/models/color_value_test.dart
@@ -3,7 +3,7 @@ import 'package:test/test.dart';
 
 void main() {
   test('ColorValue black returns proper color', () {
-    expect(ColorValue.black.value, equals('0xFF000000'));
+    expect(ColorValue.black.toHex8(), equals('0xFF000000'));
     expect(ColorValue.black.toString(), equals('const Color(0xFF000000)'));
   });
 }